fixed: VS2003 compilation warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@22848 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUITextLayout.cpp
@@ -246,7 +246,7 @@ void CGUITextLayout::BidiTransform(vector<CGUIString> &lines, bool forceLTRReadi
         sectionStyle = style;
         sectionText.clear();
       }
-      sectionText.push_back(*it & 0xffff);
+      sectionText.push_back( (DWORD)(*it & 0xffff) );
     }
 
     // handle the last section
@@ -366,7 +366,7 @@ void CGUITextLayout::ParseText(const CStdStringW &text, DWORD defaultStyle, vect
     else if (text.Mid(pos,5) == L"COLOR")
     { // color
       size_t finish = text.Find(L']', pos + 5);
-      if (on && finish != CStdString::npos && text.Find(L"[/COLOR]",finish) != CStdString::npos)
+      if (on && finish != CStdString::npos && (size_t)text.Find(L"[/COLOR]",finish) != CStdString::npos)
       { // create new color
         newColor = colors.size();
         colors.push_back(g_colorManager.GetColor(text.Mid(pos + 5, finish - pos - 5)));
